refactor: Move scalacli tests to a separate job
Previously, we would run the Scala CLI tests together with normal unit tests, but that could cause OOM issues on the CI since we would efectively start 2 Bloop instances. Now, we run the ScalaCLI tests separately.

Also:
- updated tests to run in Scala 3
- updated the version of ScalaCLI

I did those together since there is an issue with scripts alongside scala files on Scala 2, but it doesn't make sense to block on this as the version we use is only used for single scripts (which works well) and anyway the users will most likely have the Scala CLI version installed locally.
