+++ project/V.scala
@@ -34,7 +34,7 @@ object V {
   val pprint = "0.7.3"
   val sbtBloop = bloop
   val sbtJdiTools = "1.1.1"
-  val scalaCli = "0.2.1"
+  val scalaCli = "1.0.1"
   val scalafix = "0.11.0"
   val scalafmt = "3.7.4"
   val scalameta = "4.7.8"
