+++ tests/slow/src/test/scala/tests/feature/CrossCodeLensLspSuite.scala
@@ -3,7 +3,6 @@ package tests.feature
 import scala.meta.internal.metals.{BuildInfo => V}
 
 import tests.BaseCodeLensLspSuite
-import tests.ScalaCliBuildLayout
 
 class CrossCodeLensLspSuite extends BaseCodeLensLspSuite("cross-code-lens") {
 
@@ -118,32 +117,4 @@ class CrossCodeLensLspSuite extends BaseCodeLensLspSuite("cross-code-lens") {
     } yield ()
   }
 
-  test("run-main-annotation-with-script") {
-    cleanWorkspace()
-    val path = "main.sc"
-    for {
-      _ <- initialize(
-        ScalaCliBuildLayout(
-          s"""|/$path
-              |val x = 3
-              |
-              |@main def main() = {
-              |  println("annotation")
-              |}""".stripMargin,
-          V.scala3,
-        )
-      )
-      _ <- server.didOpen(path)
-      _ <- assertCodeLenses(
-        path,
-        """|<<run>><<debug>>
-           |val x = 3
-           |
-           |<<run>><<debug>>
-           |@main def main() = {
-           |  println("annotation")
-           |}""".stripMargin,
-      )
-    } yield ()
-  }
 }
