+++ tests/unit/src/test/scala/tests/CodeLensLspSuite.scala
@@ -165,52 +165,6 @@ class CodeLensLspSuite extends BaseCodeLensLspSuite("codeLenses") {
     } yield ()
   }
 
-  private val scalaCliScriptPath = "a/src/main/scala/a/main.sc"
-  test("run-script") {
-    cleanWorkspace()
-    for {
-
-      _ <- initialize(
-        s"""/.bsp/scala-cli.json
-           |${BaseScalaCliSuite.scalaCliBspJsonContent()}
-           |/.scala-build/ide-inputs.json
-           |${BaseScalaCliSuite.scalaCliIdeInputJson(".")}
-           |/$scalaCliScriptPath
-           |print("oranges are nice")""".stripMargin
-      )
-      _ <- server.didOpen(scalaCliScriptPath)
-      _ <- assertCodeLenses(
-        scalaCliScriptPath,
-        """|<<run>><<debug>>
-           |print("oranges are nice")
-           |""".stripMargin,
-      )
-    } yield ()
-  }
-
-  private val scalaCliScriptPathTop = "main.sc"
-  test("run-script-top") {
-    cleanWorkspace()
-    for {
-
-      _ <- initialize(
-        s"""/.bsp/scala-cli.json
-           |${BaseScalaCliSuite.scalaCliBspJsonContent()}
-           |/.scala-build/ide-inputs.json
-           |${BaseScalaCliSuite.scalaCliIdeInputJson(".")}
-           |/$scalaCliScriptPathTop
-           |print("oranges are nice")""".stripMargin
-      )
-      _ <- server.didOpen(scalaCliScriptPathTop)
-      _ <- assertCodeLenses(
-        scalaCliScriptPathTop,
-        """|<<run>><<debug>>
-           |print("oranges are nice")
-           |""".stripMargin,
-      )
-    } yield ()
-  }
-
   // Tests, whether main class in one project does not affect other class with same name in other project
   test("run-multi-module") {
     cleanWorkspace()
