--- tests/unit/src/test/scala/tests/ScalaCliSuite.scala
@@ -1,5 +1,0 @@
-package tests
-
-import scala.meta.internal.metals.{BuildInfo => V}
-
-class ScalaCliSuite extends tests.BaseScalaCliSuite(V.scala213)
